# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AtlasRisk HTTP API",
    "version": "1.0.0",
    "description": "Contract-first foundation for the AtlasRisk decision-support API."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/valuations/{valuation_id}/reconciliations": {
      "post": { "operationId": "createReconciliation", "parameters": [{ "name": "valuation_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReconciliationRequest" } } } }, "responses": { "201": { "description": "Immutable reconciliation checkpoint", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReconciliationCheckpoint" } } } }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
    "/api/v1/reconciliations/{reconciliation_id}": {
      "get": { "operationId": "getReconciliation", "parameters": [{ "name": "reconciliation_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Immutable reconciliation checkpoint", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReconciliationCheckpoint" } } } }, "404": { "$ref": "#/components/responses/NotFound" } } }
    }
  },
  "components": {
    "schemas": {
...
      "Problem": { "type": "object", "required": ["code", "message", "request_id"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "request_id": { "type": "string" }, "details": { "type": "object", "additionalProperties": true } } },
...
      "ReconciliationLineCheckInput": { "type": "object", "additionalProperties": false, "required": ["snapshot_line_id", "external_amount"], "properties": { "snapshot_line_id": { "type": "string", "format": "uuid" }, "external_amount": { "type": "string", "pattern": "^-?(?:0|[1-9][0-9]*)(?:\\.[0-9]+)?$" } } },
      "ReconciliationRequest": { "type": "object", "additionalProperties": false, "required": ["account_id", "source_label", "currency", "cutoff", "external_nav"], "properties": { "account_id": { "type": "string", "format": "uuid" }, "source_label": { "type": "string", "minLength": 1 }, "currency": { "type": "string", "enum": ["TRY", "USD"] }, "cutoff": { "type": "string", "format": "date-time" }, "external_nav": { "type": "string", "pattern": "^-?(?:0|[1-9][0-9]*)(?:\\.[0-9]+)?$" }, "line_checks": { "type": "array", "items": { "$ref": "#/components/schemas/ReconciliationLineCheckInput" } }, "line_checks_complete": { "type": "boolean", "default": false } } },
      "ReconciliationLineCheck": { "type": "object", "additionalProperties": false, "required": ["snapshot_line_id", "external_amount", "valuation_amount", "difference", "absolute_difference"], "properties": { "snapshot_line_id": { "type": "string", "format": "uuid" }, "external_amount": { "type": "string" }, "valuation_amount": { "type": "string" }, "difference": { "type": "string" }, "absolute_difference": { "type": "string" } } },
      "ReconciliationCheckpoint": { "type": "object", "additionalProperties": false, "required": ["id", "valuation_id", "account_id", "source_label", "currency", "cutoff", "external_nav", "valuation_nav", "absolute_difference", "effective_tolerance", "tolerance_version", "state", "line_check_state", "line_checks", "created_at"], "properties": { "id": { "type": "string", "format": "uuid" }, "valuation_id": { "type": "string", "format": "uuid" }, "account_id": { "type": "string", "format": "uuid" }, "source_label": { "type": "string" }, "currency": { "type": "string", "enum": ["TRY", "USD"] }, "cutoff": { "type": "string", "format": "date-time" }, "external_nav": { "type": "string" }, "valuation_nav": { "type": "string" }, "absolute_difference": { "type": "string" }, "relative_difference": { "type": ["string", "null"] }, "effective_tolerance": { "type": "string" }, "tolerance_version": { "type": "integer", "minimum": 0 }, "state": { "type": "string", "enum": ["reconciled", "unreconciled"] }, "reason_code": { "type": "string" }, "line_check_state": { "type": "string", "enum": ["none", "partial", "complete"] }, "line_checks": { "type": "array", "items": { "$ref": "#/components/schemas/ReconciliationLineCheck" } }, "created_at": { "type": "string", "format": "date-time" } } },
...
    },
...
    "responses": {
      "InvalidRequest": { "description": "Invalid series identifier, mode, limit, cursor, or time window", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Conflict": { "description": "Resource conflicts with an existing identifier or immutable reference", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
    },
...
  }
}